Move coordinate-based structs to a new module
