+++ common/src/lib.rs
@@ -29,6 +29,7 @@ pub mod proto;
 mod sim_config;
 pub mod terraingen;
 pub mod traversal;
+pub mod voxel_math;
 pub mod world;
 pub mod worldgen;
