+++ common/src/proto.rs
@@ -1,11 +1,8 @@
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    dodeca,
-    graph::NodeId,
-    node::{ChunkId, Coords},
-    world::Material,
-    EntityId, SimConfig, Step,
+    dodeca, graph::NodeId, node::ChunkId, voxel_math::Coords, world::Material, EntityId, SimConfig,
+    Step,
 };
 
 #[derive(Debug, Serialize, Deserialize)]
